refactor: перенесла shared
