+++ 10-weather-typescript/services/shared/helpers.ts
