+++ 10-weather-typescript/services/shared/httpUtils.ts
